+++ src/main/java/com/blackbuild/groovy/configdsl/transform/Default.java
@@ -8,16 +8,26 @@ import java.lang.annotation.Target;
 /**
  * Designates a default value for the given field. This automatically creates a getter providing that default value.
  *
- * Can be used in one of two modes. The simple mode, using the value of the annotation delegates simply to another field,
- * the complex one provides a closure for the default value.
+ * The default value can be one of the following (as defined by which attribute of the annotation is set, which is used
+ * if the current value is not set (as defined by Groovy Truth):
+ * - field: return the value of the target field
+ * - delegate: return the value of an identically named field of the given delegate field
+ * - closure: execute the closure (in the context of `this` and return the result
+ *
+ * Note that it is illegal to set more than one member of the `Default` annotation.
  */
 @Target(ElementType.FIELD)
 @Retention(RetentionPolicy.CLASS)
 public @interface Default {
 
+    @Deprecated
     String value() default "";
 
+    String field() default "";
+
     Class code() default None.class;
+
+    String delegate() default "";
 
     class None {}
 }