+++ src/main/java/com/blackbuild/groovy/configdsl/transform/ast/DefaultMethods.java
@@ -5,16 +5,14 @@ import org.codehaus.groovy.ast.AnnotationNode;
 import org.codehaus.groovy.ast.ClassHelper;
 import org.codehaus.groovy.ast.ClassNode;
 import org.codehaus.groovy.ast.FieldNode;
-import org.codehaus.groovy.ast.expr.ClassExpression;
-import org.codehaus.groovy.ast.expr.ClosureExpression;
-import org.codehaus.groovy.ast.expr.ElvisOperatorExpression;
-import org.codehaus.groovy.ast.expr.Expression;
+import org.codehaus.groovy.ast.expr.*;
 import org.jetbrains.annotations.Nullable;
 
 import static com.blackbuild.groovy.configdsl.transform.ast.ASTHelper.getAnnotation;
 import static com.blackbuild.groovy.configdsl.transform.ast.MethodBuilder.createPublicMethod;
 import static java.lang.Character.toUpperCase;
 import static org.codehaus.groovy.ast.ClassHelper.make;
+import static org.codehaus.groovy.ast.expr.CastExpression.asExpression;
 import static org.codehaus.groovy.ast.tools.GeneralUtils.*;
 import static org.codehaus.groovy.transform.AbstractASTTransformation.getMemberStringValue;
 
@@ -23,6 +21,7 @@ import static org.codehaus.groovy.transform.AbstractASTTransformation.getMemberS
  * Helper class for default values.
  */
 public class DefaultMethods {
+    public static final String CLOSURE_VAR_NAME = "closure";
     private DSLASTTransformation transformation;
 
     static final ClassNode DEFAULT_ANNOTATION = make(Default.class);
@@ -44,30 +43,28 @@ public class DefaultMethods {
     }
 
     private void createDefaultValueFor(FieldNode fieldNode, AnnotationNode defaultAnnotation) {
+        assertOnlyOneMemberOfAnnotationIsSet(defaultAnnotation);
 
-        String delegate = getMemberStringValue(defaultAnnotation, "value", "");
-        ClosureExpression code = getCodeClosureFor(fieldNode, defaultAnnotation);
-        assertOnlyDelegateOrCodeClosureIsSet(fieldNode, delegate, code);
+        String fieldMember = getMemberStringValue(defaultAnnotation, "value");
 
-        if (!delegate.isEmpty())
-            createDelegateMethod(fieldNode, delegate);
-        else
+        if (fieldMember == null)
+            fieldMember = getMemberStringValue(defaultAnnotation, "field");
+
+        if (fieldMember != null) {
+            createFieldMethod(fieldNode, fieldMember);
+            return;
+        }
+
+        ClosureExpression code = getCodeClosureFor(defaultAnnotation);
+        if (code != null) {
             createClosureMethod(fieldNode, code);
-    }
+            return;
+        }
 
-    private void createClosureMethod(FieldNode fieldNode, ClosureExpression code) {
-        String ownGetter = getGetterName(fieldNode.getName());
-
-        createPublicMethod(ownGetter)
-                .returning(fieldNode.getType())
-                .declareVariable("closure", code)
-                .assignS(propX(varX("closure"), "delegate"), varX("this"))
-                .assignS(
-                        propX(varX("closure"), "resolveStrategy"),
-                        propX(classX(ClassHelper.CLOSURE_TYPE), "DELEGATE_FIRST")
-                )
-                .statement(new ElvisOperatorExpression(varX(fieldNode.getName()), callX(varX("closure"), "call")))
-                .addTo(transformation.annotatedClass);
+        String delegateMember = getMemberStringValue(defaultAnnotation, "delegate");
+        if (delegateMember != null) {
+            createDelegateMethod(fieldNode, delegateMember);
+        }
     }
 
     private void createDelegateMethod(FieldNode fieldNode, String delegate) {
@@ -76,7 +73,32 @@ public class DefaultMethods {
 
         createPublicMethod(ownGetter)
                 .returning(fieldNode.getType())
-                .statement(new ElvisOperatorExpression(varX(fieldNode.getName()), callThisX(delegateGetter)))
+                .statement(asExpression(fieldNode.getType(), new ElvisOperatorExpression(varX(fieldNode.getName()), new PropertyExpression(callThisX(delegateGetter), new ConstantExpression(fieldNode.getName()), true))))
+                .addTo(transformation.annotatedClass);
+    }
+
+    private void createClosureMethod(FieldNode fieldNode, ClosureExpression code) {
+        String ownGetter = getGetterName(fieldNode.getName());
+
+        createPublicMethod(ownGetter)
+                .returning(fieldNode.getType())
+                .declareVariable(CLOSURE_VAR_NAME, code)
+                .assignS(propX(varX(CLOSURE_VAR_NAME), "delegate"), varX("this"))
+                .assignS(
+                        propX(varX(CLOSURE_VAR_NAME), "resolveStrategy"),
+                        propX(classX(ClassHelper.CLOSURE_TYPE), "DELEGATE_FIRST")
+                )
+                .statement(asExpression(fieldNode.getType(), new ElvisOperatorExpression(varX(fieldNode.getName()), callX(varX(CLOSURE_VAR_NAME), "call"))))
+                .addTo(transformation.annotatedClass);
+    }
+
+    private void createFieldMethod(FieldNode fieldNode, String targetField) {
+        String ownGetter = getGetterName(fieldNode.getName());
+        String fieldGetter = getGetterName(targetField);
+
+        createPublicMethod(ownGetter)
+                .returning(fieldNode.getType())
+                .statement(asExpression(fieldNode.getType(), new ElvisOperatorExpression(varX(fieldNode.getName()), callThisX(fieldGetter))))
                 .addTo(transformation.annotatedClass);
     }
 
@@ -85,26 +107,26 @@ public class DefaultMethods {
     }
 
 
-    private void assertOnlyDelegateOrCodeClosureIsSet(FieldNode fieldNode, String delegate, ClosureExpression code) {
-        if (!delegate.isEmpty() && code != null)
-            transformation.addError("Only either a default property or a closure is allowed, not both!", fieldNode);
+    private void assertOnlyOneMemberOfAnnotationIsSet(AnnotationNode annotationNode) {
+        int numberOfMembers = annotationNode.getMembers().size();
+
+        if (numberOfMembers == 0)
+            transformation.addError("You must define either delegate, code or field for @Default annotations", annotationNode);
+
+        if (numberOfMembers > 1)
+            transformation.addError("Only one member for @Default annotation is allowed!", annotationNode);
     }
 
     @Nullable
-    private ClosureExpression getCodeClosureFor(FieldNode fieldNode, AnnotationNode defaultAnnotation) {
+    private ClosureExpression getCodeClosureFor(AnnotationNode defaultAnnotation) {
         Expression codeExpression = defaultAnnotation.getMember("code");
-        ClosureExpression code;
         if (codeExpression == null)
             return null;
-        if (codeExpression instanceof ClassExpression) {
-            code = null;
-        } else if (codeExpression instanceof ClosureExpression){
-            code = (ClosureExpression) codeExpression;
-        } else {
-            transformation.addError("Illegal value for code, only None.class or a closure is allowed.", fieldNode);
-            code = null;
-        }
-        return code;
+        if (codeExpression instanceof ClosureExpression)
+            return (ClosureExpression) codeExpression;
+
+        transformation.addError("Illegal value for code, only None.class or a closure is allowed.", defaultAnnotation);
+        return null;
     }