+++ CHANGES.md
@@ -4,6 +4,7 @@
 - new `makeTemplate` method, which only creates template, without assigning it
 - [Lifecycle methods](https://github.com/blackbuild/config-dsl#Lifecycle-Methods) can be created using the `@PostCreate` and `@PostApply` annotations (see [#38](https://github.com/blackbuild/config-dsl/issues/38))
 - DSL classes are now `Serializable` (see [#35](https://github.com/blackbuild/config-dsl/issues/35)) 
+- `@Default` now also supports `delegate` members (see [#46](https://github.com/blackbuild/config-dsl/issues/46))
 
 ## 0.20
 - A new, explicit [Template Mechanism](https://github.com/blackbuild/config-dsl#Template-Mechanism) (see [#34](https://github.com/blackbuild/config-dsl/issues/34))